requirements.txt
absl-py==1.4.0
albumentations==1.3.1
astunparse==1.6.3
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==2.0.4
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.6.1
gast==0.4.0
gmpy2==2.1.2
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
jax==0.4.9
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.3
matplotlib==3.9.0
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pip==24.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.32.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.6.18
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.48
ultralytics-thop==2.0.0
urllib3==2.2.1
wheel==0.43.0
win-inet-pton==1.1.0
zipp==3.19.2
